+++ src/main/java/com/github/wintersteve25/tau/components/animated/AnimatedTexture.java
@@ -7,59 +7,52 @@ import com.github.wintersteve25.tau.layout.Layout;
 import com.github.wintersteve25.tau.theme.Theme;
 import com.github.wintersteve25.tau.utils.SimpleVec2i;
 import com.github.wintersteve25.tau.utils.Size;
+import com.github.wintersteve25.tau.utils.Variable;
 import net.minecraft.resources.ResourceLocation;
-
-import java.util.function.Supplier;
 
 public class AnimatedTexture implements PrimitiveUIComponent {
 
     private final ResourceLocation textureLocation;
     private final SimpleVec2i textureSize;
-    private final SimpleVec2i uvSize;
-    private final SimpleVec2i size;
-    private final Supplier<Size> uvSizeAnimator;
-    private final Supplier<SimpleVec2i> uvPosAnimator;
+    private final Size size;
+    private final Variable<SimpleVec2i> uvSize;
+    private final Variable<SimpleVec2i> uvPos;
 
-    public AnimatedTexture(ResourceLocation textureLocation, SimpleVec2i textureSize, SimpleVec2i uvSize, SimpleVec2i size, Supplier<Size> uvSizeAnimator, Supplier<SimpleVec2i> uvPosAnimator) {
+    public AnimatedTexture(ResourceLocation textureLocation, SimpleVec2i textureSize, Size size, Variable<SimpleVec2i> uvSize, Variable<SimpleVec2i> uvPos) {
         this.textureLocation = textureLocation;
         this.textureSize = textureSize;
-        this.uvSize = uvSize;
         this.size = size;
-        this.uvSizeAnimator = uvSizeAnimator;
-        this.uvPosAnimator = uvPosAnimator;
+        this.uvSize = uvSize;
+        this.uvPos = uvPos;
     }
 
     @Override
     public SimpleVec2i build(Layout layout, Theme theme, BuildContext context) {
-
-        SimpleVec2i position = layout.getPosition(uvSize);
+        SimpleVec2i actualSize = size.get(layout.getSize());
+        SimpleVec2i position = layout.getPosition(actualSize);
 
         context.renderables().add((graphics, pMouseX, pMouseY, pPartialTicks) -> {
-            SimpleVec2i newUvSize = new SimpleVec2i(uvSize.x, uvSize.y);
-            newUvSize = uvSizeAnimator.get().get(newUvSize);
-            SimpleVec2i newUvPos = uvPosAnimator.get();
-
-            graphics.blit(textureLocation, position.x, position.y, size.x, size.y, newUvPos.x, newUvPos.y, newUvSize.x, newUvSize.y,
+            graphics.blit(textureLocation, position.x, position.y, actualSize.x, actualSize.y, uvPos.getValue().x, uvPos.getValue().y, uvSize.getValue().x, uvSize.getValue().y,
                     textureSize.x,
                     textureSize.y);
         });
 
-        return size;
+        return actualSize;
     }
 
     public static final class Builder implements UIComponent {
 
         private final ResourceLocation textureLocation;
-        private final Supplier<Size> uvSizeAnimator;
-        private final Supplier<SimpleVec2i> uvPosAnimator;
+        private final Variable<SimpleVec2i> uvSize;
+        private final Variable<SimpleVec2i> uvPos;
+        private final Size size;
         private SimpleVec2i textureSize;
-        private SimpleVec2i uvSize;
-        private SimpleVec2i size;
 
-        public Builder(ResourceLocation textureLocation, Supplier<Size> uvSizeAnimator, Supplier<SimpleVec2i> uvPosAnimator) {
+        public Builder(ResourceLocation textureLocation, Size size, Variable<SimpleVec2i> uvSize, Variable<SimpleVec2i> uvPos) {
             this.textureLocation = textureLocation;
-            this.uvSizeAnimator = uvSizeAnimator;
-            this.uvPosAnimator = uvPosAnimator;
+            this.uvSize = uvSize;
+            this.uvPos = uvPos;
+            this.size = size;
         }
 
         public Builder withTextureSize(SimpleVec2i textureSize) {
@@ -67,20 +60,9 @@ public class AnimatedTexture implements PrimitiveUIComponent {
             return this;
         }
 
-        public Builder withUvSize(SimpleVec2i uvSize) {
-            this.uvSize = uvSize;
-            return this;
-        }
-
-        public Builder withSize(SimpleVec2i size) {
-            this.size = size;
-            return this;
-        }
-
         public UIComponent build() {
             textureSize = textureSize == null ? new SimpleVec2i(256, 256) : textureSize;
-            uvSize = uvSize == null ? textureSize : uvSize;
-            return new AnimatedTexture(textureLocation, textureSize, uvSize, size == null ? uvSize : size, uvSizeAnimator, uvPosAnimator);
+            return new AnimatedTexture(textureLocation, textureSize, size, uvSize, uvPos);
         }
 
         @Override